SEO Project Task/seo_blog_generator/requirements.txt
openai==0.10.0
gunicorn==21.2.0
Jinja2==3.1.3
Werkzeug==3.0.1
scikit-learn==1.0.2
Flask-Migrate==4.0.7
matplotlib==3.4.3
WTForms==3.1.2
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.2.1
numpy==1.26.4
pandas==2.2.1
markdown==3.4.1
sqlite==3.39.2



Step - 1
Create a Virtual environment

Step - 2
Create a Project folder and app

Step - 3
Install the requirements
pip install -r requirements.txt


Step - 4
Create a HTML and CSS files(templates and static folder)

Step - 5
Run Database Migrations

Step - 6
Start the Application

Step - 7
Open in Browser
Visit http://127.0.0.1:5000/ to access the blog. 